.pub-cache/hosted/pub.dartlang.org/html-0.15.2/test/data/tree-construction/adoption02.dat

#data
<b>1<i>2<p>3</b>4
#errors
#document
| <html>
|   <head>
|   <body>
|     <b>
|       "1"
|       <i>
|         "2"
|     <i>
|       <p>
|         <b>
|           "3"
|         "4"

#data
<a><div><style></style><address><a>
#errors
#document
| <html>
|   <head>
|   <body>
|     <a>
|     <div>
|       <a>
|         <style>
|       <address>
|         <a>
|         <a>
